add TODO for firestore
