+++ lib/pages/play/class/matches_container.dart
@@ -1,27 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:pocket_eleven/design/colors.dart';
 
-class StuffView extends StatelessWidget {
-  const StuffView({super.key});
+class MatchesContainer extends StatelessWidget {
+  const MatchesContainer({
+    required this.screenWidth,
+    super.key,
+  });
+
+  final double screenWidth;
 
   @override
   Widget build(BuildContext context) {
-    final double screenWidth = MediaQuery.of(context).size.width;
-    final double screenHeight = MediaQuery.of(context).size.height;
-
     return Container(
-      margin: EdgeInsets.all(screenWidth * 0.04),
+      margin: EdgeInsets.all(screenWidth * 0.05),
+      padding: EdgeInsets.all(screenWidth * 0.04),
       decoration: BoxDecoration(
         color: AppColors.hoverColor,
         border: Border.all(color: AppColors.borderColor, width: 1),
         borderRadius: BorderRadius.circular(10.0),
       ),
-      width: screenWidth,
-      height: screenHeight,
       child: const Center(
+        // TODO: Implement upcoming matches
         child: Text(
-          // TODO: Implement Stuff transfers
-          'Stuff Container',
+          'Matches Container',
           style: TextStyle(
             color: AppColors.textEnabledColor,
             fontSize: 18,
